2025_7_b passing and correct.
Solved the rather challenging problem by instead of keeping track of each new beam, keeping track of how many beams in each index. There is no need to simulate hundres/thousands of the same beams in a single index getting split.

Still a bit of work to be done. Want to make everything index into an array as we know the total max size.